tfs_dav_filter: replace references to DWORD with ULONG, DWORD does not exist in kernel builds
--HG--
extra : rebase_source : 2a900dc33b191ff96c28081870dad6da27d40d52

+++ dependencies/tfs_dav_filter/ramdisk_device.hpp
@@ -73,7 +73,7 @@ class RAMDiskDevice : public IODevice {
   PnPState pnp_state;
 
   PDEVICE_OBJECT control_device;
-  DWORD engage_count;
+  ULONG engage_count;
 
   SIZE_T
   get_image_size() const noexcept;
